refactor initialization so we only load tinymce if a d2l-html-editor is included on the page
