+++ demo/index.html
@@ -13,7 +13,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 	<meta charset="utf-8">
 	<meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
 	<title>d2l-html-editor Demo</title>
-	<script src="https://s.brightspace.com/lib/tinymce/dev/4.8.5-a11ychecker.1.2.1-53-powerpaste.3.3.3-308-shadow-dom-fork-3/tinymce.js"></script>
+	<!-- <script src="https://s.brightspace.com/lib/tinymce/dev/4.8.5-a11ychecker.1.2.1-53-powerpaste.3.3.3-308-shadow-dom-fork-3/tinymce.js"></script> -->
 	<script src="../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
 	<script type="module" src="./d2l-editor-wrapper.js"></script>
